requirements.txt
arrow==0.17.0
asgiref==3.3.1
binaryornot==0.4.4
cairocffi==1.2.0
CairoSVG==2.5.0
certifi==2020.12.5
cffi==1.14.4
chardet==4.0.0
click==7.1.2
cookiecutter==1.7.2
cssselect2==0.4.1
defusedxml==0.6.0
Django==3.1.4
django-debug-toolbar==3.2
django-environ==0.4.5
django-extensions==3.1.0
django-grappelli==2.14.3
django-jquery==3.1.0
environ==1.0
html5lib==1.1
idna==2.10
Jinja2==2.11.2
jinja2-time==0.2.0
MarkupSafe==1.1.1
mod-wsgi==4.7.1
Pillow==8.0.1
poyo==0.5.0
psycopg2==2.8.6
pycparser==2.20
Pyphen==0.10.0
python-dateutil==2.8.1
python-slugify==4.0.1
pytz==2020.5
requests==2.25.1
six==1.15.0
sqlparse==0.4.1
text-unidecode==1.3
tinycss2==1.1.0
urllib3==1.26.2
WeasyPrint==52.2
webencodings==0.5.1
